exams/2024/06-13/solution/pyproject.toml
[tool.poetry]
name = "internet-programming-104779-exam-2024-06-13"
version = "0.1.0"
description = "Yet Another Task Management System"
authors = ["Alex Carrega <alessandro.carrega@unige.it>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bcrypt = "^4.1.3"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
PyJWT = "^2.8.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
